[module5-task1] implement PageController
